requirements.txt
groq
streamlit
PyGithub
plotly
pandas
matplotlib
python-dotenv